trying something new for elytra dying
figured out how to make elytras fully dyeable like leather armor, and it seems to work well! now there's a new system that will also allow the resource pack to be way simpler, and it should make it easier if I want to add more variants in the future.

The problem with this new system is that any dyed elytras from the original versions of the pack use item models in not exactly the best way, which idealy I would like to remove. I need to figure out a way to grandfather them in without messing them up, but I think I'm gonna need to keep the old sprites in for a while longer at the very least. gonna be working on that